Feat(bungie.net) requests to twab successful
The bungie request code is now working properly, and returns the most recent twab or twid. Now I will work on integrating bungie's twitter into the code.
