--- HELMET/index.html
@@ -1,8 +1,0 @@
-<html>
-    <body>
-        <h1>
-            H.E.L.M.E.T
-        </h1>
-        <iframe src="https://q4kk.github.io/#this-is-a-test" frameborder="0.25"></iframe>
-    </body>
-</html>
